Début de la mise en place d'une gestion des versions du programme et de la base. MAJ de class.DB (bug php 5.3).
git-svn-id: https://svn.devsesamath.net/svn/sacoche/trunk@72 3af1deee-4ace-4d6f-927a-55d690a2d960

+++ _inc/fonction_clean.php
@@ -193,7 +193,7 @@ function vide_dossier($dossier,$nb_minutes)
 	{
 		$extension = pathinfo($fichier_nom,PATHINFO_EXTENSION);
 		$date_unix = filemtime($dossier.$fichier_nom);
-		if( ($date_unix<$date_limite) && ($extension!='htm') )
+		if( (is_file($dossier.$fichier_nom)) && ($date_unix<$date_limite) && ($extension!='htm') )
 		{
 			unlink($dossier.$fichier_nom);
 		}
